Revert "Start renames, add RequestConverters"
This reverts commit 48f9514bfa2dd0a0766140d4d4d8dc1d09452943.
