+++ x-pack/plugin/index-lifecycle/src/main/java/org/elasticsearch/xpack/indexlifecycle/action/TransportSetPolicyForIndexAction.java
@@ -23,21 +23,21 @@ import org.elasticsearch.threadpool.ThreadPool;
 import org.elasticsearch.transport.TransportService;
 import org.elasticsearch.xpack.core.indexlifecycle.IndexLifecycleMetadata;
 import org.elasticsearch.xpack.core.indexlifecycle.LifecyclePolicy;
-import org.elasticsearch.xpack.core.indexlifecycle.action.SetIndexPolicyAction;
-import org.elasticsearch.protocol.xpack.indexlifecycle.SetIndexPolicyRequest;
-import org.elasticsearch.protocol.xpack.indexlifecycle.SetIndexPolicyResponse;
+import org.elasticsearch.xpack.core.indexlifecycle.action.SetPolicyForIndexAction;
+import org.elasticsearch.xpack.core.indexlifecycle.action.SetPolicyForIndexAction.Request;
+import org.elasticsearch.xpack.core.indexlifecycle.action.SetPolicyForIndexAction.Response;
 import org.elasticsearch.xpack.indexlifecycle.IndexLifecycleRunner;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class TransportSetIndexPolicyAction extends TransportMasterNodeAction<SetIndexPolicyRequest, SetIndexPolicyResponse> {
+public class TransportSetPolicyForIndexAction extends TransportMasterNodeAction<Request, Response> {
 
     @Inject
-    public TransportSetIndexPolicyAction(Settings settings, TransportService transportService, ClusterService clusterService,
-                                         ThreadPool threadPool, ActionFilters actionFilters, IndexNameExpressionResolver indexNameExpressionResolver) {
-        super(settings, SetIndexPolicyAction.NAME, transportService, clusterService, threadPool, actionFilters,
-                indexNameExpressionResolver, SetIndexPolicyRequest::new);
+    public TransportSetPolicyForIndexAction(Settings settings, TransportService transportService, ClusterService clusterService,
+            ThreadPool threadPool, ActionFilters actionFilters, IndexNameExpressionResolver indexNameExpressionResolver) {
+        super(settings, SetPolicyForIndexAction.NAME, transportService, clusterService, threadPool, actionFilters,
+                indexNameExpressionResolver, Request::new);
     }
 
     @Override
@@ -46,22 +46,21 @@ public class TransportSetIndexPolicyAction extends TransportMasterNodeAction<Set
     }
 
     @Override
-    protected SetIndexPolicyResponse newResponse() {
-        return new SetIndexPolicyResponse();
+    protected Response newResponse() {
+        return new Response();
     }
 
     @Override
-    protected ClusterBlockException checkBlock(SetIndexPolicyRequest request, ClusterState state) {
+    protected ClusterBlockException checkBlock(Request request, ClusterState state) {
         return state.blocks().globalBlockedException(ClusterBlockLevel.METADATA_WRITE);
     }
 
     @Override
-    protected void masterOperation(SetIndexPolicyRequest request, ClusterState state,
-                                   ActionListener<SetIndexPolicyResponse> listener) throws Exception {
+    protected void masterOperation(Request request, ClusterState state, ActionListener<Response> listener) throws Exception {
         final String newPolicyName = request.policy();
         final Index[] indices = indexNameExpressionResolver.concreteIndices(state, request.indicesOptions(), request.indices());
         clusterService.submitStateUpdateTask("change-lifecycle-for-index-" + newPolicyName,
-                new AckedClusterStateUpdateTask<SetIndexPolicyResponse>(request, listener) {
+                new AckedClusterStateUpdateTask<Response>(request, listener) {
 
                     private final List<String> failedIndexes = new ArrayList<>();
 
@@ -89,8 +88,8 @@ public class TransportSetIndexPolicyAction extends TransportMasterNodeAction<Set
                     }
 
                     @Override
-                    protected SetIndexPolicyResponse newResponse(boolean acknowledged) {
-                        return new SetIndexPolicyResponse(failedIndexes);
+                    protected Response newResponse(boolean acknowledged) {
+                        return new Response(failedIndexes);
                     }
                 });
     }
